refactor: streamline string formatting in test files
- Replaced `format!("{}/api/auth/...")` with `format!("{base_url}/api/auth/...")` for consistency across multiple test files.
- Updated assertions to use more concise formatting for error messages, enhancing readability.
- Refactored character checks in `password_reset_token.rs` to utilize `char::is_alphanumeric` for improved clarity.

These changes aim to enhance code consistency and readability in the test suite.

+++ IAMRusty/domain/src/service/token_service.rs
@@ -174,7 +174,7 @@ mod tests {
                 .withf(move |claims: &TokenClaims| {
                     claims.sub == user_id_clone
                         && claims.username == username_clone
-                        && claims.exp > Utc::now().timestamp() as i64 // Token should expire in the future
+                        && claims.exp > Utc::now().timestamp() // Token should expire in the future
                 })
                 .returning(|_| Ok("token".to_string()));
